refactor(codec): Overhaul parser with robust error handling and bytes-first processing
